@@ -1,6 +1,6 @@
 import {Inject, Injectable} from '@angular/core';
 import {Http, Response} from "@angular/http";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs/Rx";
 import {Containers} from "../../store/data/containers/container.model";
 import NewContainer = Containers.NewContainer;
 
