@@ -1,6 +1,6 @@
 import {Inject, Injectable} from '@angular/core';
 import {Http} from "@angular/http";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs/Rx";
 
 @Injectable()
 export class SummariesService {
