@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs/Rx';
 import { PackageReport } from './package-report.model';
 import { Vulnerability } from './vulnerability.model';
 @Injectable()
