@@ -1,15 +1,10 @@
 import {AfterViewInit, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {ContainersService} from "../../services/containers.service";
-import {
-  IDatatablePaginationEvent,
-  IDatatableSelectionEvent, IDatatableSortEvent, MdDataTableComponent,
-  MdDataTablePaginationComponent
-} from "ng2-md-datatable";
 import {Observable} from "rxjs/Observable";
 import 'rxjs/add/observable/from';
 import {Subject} from "rxjs/Subject";
 import {NewContainerModalComponent} from "./new/new.container.modal.component";
-import {MdDialog} from "@angular/material";
+import {MatDialog} from "@angular/material";
 
 @Component({
   selector: 'app-containers',
@@ -17,9 +12,6 @@ import {MdDialog} from "@angular/material";
   styleUrls: ['./containers.component.less']
 })
 export class ContainersComponent implements OnInit, AfterViewInit, OnDestroy {
-
-  @ViewChild(MdDataTableComponent) datatable: MdDataTableComponent;
-  @ViewChild(MdDataTablePaginationComponent) pager: MdDataTablePaginationComponent;
 
   public pagination: any = {
     Pages: 0,
@@ -30,7 +22,7 @@ export class ContainersComponent implements OnInit, AfterViewInit, OnDestroy {
   private unmount$: Subject<void> = new Subject<void>();
 
   constructor(private containersService: ContainersService,
-              public dialog: MdDialog) { }
+              public dialog: MatDialog) { }
 
   ngOnInit() {
 
@@ -46,21 +38,7 @@ export class ContainersComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   ngAfterViewInit() {
-    if (this.datatable) {
-      Observable.from(this.datatable.selectionChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatableSelectionEvent) => console.log("Data selected"));
 
-      Observable.from(this.datatable.sortChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatableSortEvent) =>
-          console.log("Sort order"));
-
-      Observable.from(this.pager.paginationChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatablePaginationEvent) =>
-          console.log("Pager changed"));
-    }
   }
 
   openDialog() {
