@@ -1,15 +1,10 @@
 import {AfterViewInit, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {ContainersService} from "../../../services/containers.service";
-import {
-  IDatatablePaginationEvent,
-  IDatatableSelectionEvent, IDatatableSortEvent, MdDataTableComponent,
-  MdDataTablePaginationComponent
-} from "ng2-md-datatable";
 import {Observable} from "rxjs/Observable";
 import 'rxjs/add/observable/from';
 import {Subject} from "rxjs/Subject";
 import {RegistryNewContainerModalComponent} from "./new/registry.new.container.modal.component";
-import {MdDialog} from "@angular/material";
+import {MatDialog} from "@angular/material";
 import {RegistriesService} from "../../../services/registries.service";
 import {ActivatedRoute, Route} from "@angular/router";
 
@@ -19,9 +14,6 @@ import {ActivatedRoute, Route} from "@angular/router";
   styleUrls: ['./registry.containers.component.less']
 })
 export class RegistryContainersComponent implements OnInit, AfterViewInit, OnDestroy {
-
-  @ViewChild(MdDataTableComponent) datatable: MdDataTableComponent;
-  @ViewChild(MdDataTablePaginationComponent) pager: MdDataTablePaginationComponent;
 
   private id: number;
 
@@ -36,7 +28,7 @@ export class RegistryContainersComponent implements OnInit, AfterViewInit, OnDes
   constructor(private containersService: ContainersService,
               private registriesService: RegistriesService,
               private route: ActivatedRoute,
-              public dialog: MdDialog) { }
+              public dialog: MatDialog) { }
 
   ngOnInit() {
 
@@ -55,21 +47,6 @@ export class RegistryContainersComponent implements OnInit, AfterViewInit, OnDes
   }
 
   ngAfterViewInit() {
-    if (this.datatable) {
-      Observable.from(this.datatable.selectionChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatableSelectionEvent) => console.log("Data selected"));
-
-      Observable.from(this.datatable.sortChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatableSortEvent) =>
-          console.log("Sort order"));
-
-      Observable.from(this.pager.paginationChange)
-        .takeUntil(this.unmount$)
-        .subscribe((e: IDatatablePaginationEvent) =>
-          console.log("Pager changed"));
-    }
   }
 
   openDialog() {
