@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { MdIconRegistry } from '@angular/material';
+import { MatIconRegistry } from '@angular/material';
 import {DomSanitizer} from "@angular/platform-browser";
 
 @Component({
@@ -11,7 +11,7 @@ export class HeaderComponent implements OnInit {
 
   private selectedRegistry: any;
 
-  constructor(iconRegistry: MdIconRegistry, sanitizer: DomSanitizer) {
+  constructor(iconRegistry: MatIconRegistry, sanitizer: DomSanitizer) {
     iconRegistry.addSvgIcon(
       'github-circle',
         sanitizer.bypassSecurityTrustResourceUrl('assets/icons/github-circle.svg'));
