@@ -2,62 +2,67 @@
  * Created by kevinbayes on 13/04/17.
  */
 import {
-  MdButtonModule,
-  MdCardModule,
-  MdCheckboxModule,
-  MdGridListModule,
-  MdIconModule,
-  MdInputModule,
-  MdListModule,
-  MdMenuModule,
-  MdProgressSpinnerModule,
-  MdRippleModule,
-  MdSidenavModule,
-  MdSnackBarModule,
-  MdTabsModule,
-  MdToolbarModule,
-  MdDialogModule, MdSelectModule, MdChipsModule,
+  MatButtonModule,
+  MatCardModule,
+  MatCheckboxModule,
+  MatGridListModule,
+  MatIconModule,
+  MatInputModule,
+  MatListModule,
+  MatMenuModule,
+  MatProgressSpinnerModule,
+  MatRippleModule,
+  MatSidenavModule,
+  MatSnackBarModule,
+  MatTabsModule,
+  MatToolbarModule,
+  MatDialogModule,
+  MatSelectModule,
+  MatChipsModule,
+  MatTableModule
 } from '@angular/material';
 import { NgModule } from "@angular/core";
 
 @NgModule({
   imports:      [
-    MdButtonModule,
-    MdCheckboxModule,
-    MdToolbarModule,
-    MdCardModule,
-    MdMenuModule,
-    MdIconModule,
-    MdInputModule,
-    MdGridListModule,
-    MdListModule,
-    MdProgressSpinnerModule,
-    MdSnackBarModule,
-    MdRippleModule,
-    MdSidenavModule,
-    MdTabsModule,
-    MdDialogModule,
-    MdSelectModule,
-    MdChipsModule,
+    MatButtonModule,
+    MatCardModule,
+    MatCheckboxModule,
+    MatGridListModule,
+    MatIconModule,
+    MatInputModule,
+    MatListModule,
+    MatMenuModule,
+    MatProgressSpinnerModule,
+    MatRippleModule,
+    MatSidenavModule,
+    MatSnackBarModule,
+    MatTabsModule,
+    MatToolbarModule,
+    MatDialogModule,
+    MatSelectModule,
+    MatChipsModule,
+    MatTableModule
   ],
   exports:      [
-    MdButtonModule,
-    MdCheckboxModule,
-    MdToolbarModule,
-    MdCardModule,
-    MdMenuModule,
-    MdIconModule,
-    MdInputModule,
-    MdGridListModule,
-    MdListModule,
-    MdProgressSpinnerModule,
-    MdSnackBarModule,
-    MdRippleModule,
-    MdSidenavModule,
-    MdTabsModule,
-    MdDialogModule,
-    MdSelectModule,
-    MdChipsModule,
+    MatButtonModule,
+    MatCardModule,
+    MatCheckboxModule,
+    MatGridListModule,
+    MatIconModule,
+    MatInputModule,
+    MatListModule,
+    MatMenuModule,
+    MatProgressSpinnerModule,
+    MatRippleModule,
+    MatSidenavModule,
+    MatSnackBarModule,
+    MatTabsModule,
+    MatToolbarModule,
+    MatDialogModule,
+    MatSelectModule,
+    MatChipsModule,
+    MatTableModule
   ]
 })
 export class MaterialModule { }
